restructured past workshops content
