+++ src/sections/PastWorkshopsSection.js
@@ -1,153 +1,84 @@
-import { useState, useEffect } from 'react';
-import WorkshopWindow from '../components/WorkshopWindow';
-import styles from '../styles/sections/PastWorkshopsSection.module.css';
+import React, { useState } from 'react';
+import { FaChevronLeft, FaChevronRight } from 'react-icons/fa';
+import styles from '@/styles/sections/PastWorkshopsSection.module.css';
+import WorkshopSeries from './WorkshopSeries';
 
-export default function PastWorkshopsSection() {
-  const data = [
-    {
-      'workshop series': 'Cloud Computing',
-      workshops: [
-        {
-          title: 'Cloud Computing 1',
-          description: 'Description',
-          emoji: '/assets/design-vectors/cloud-computing.svg',
-          slides:
-            'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-          code: 'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-          recording:
-            'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-        },
-        {
-          title: 'Cloud Computing 2',
-          description: 'Description',
-          emoji: '/assets/design-vectors/cloud-computing.svg',
-          slides:
-            'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-          code: 'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-        },
-        {
-          title: 'Cloud Computing 3',
-          description: 'Description',
-          emoji: '/assets/design-vectors/cloud-computing.svg',
-          slides:
-            'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-          code: 'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-        },
-        {
-          title: 'Cloud Computing 4',
-          description: 'Description',
-          emoji: '/assets/design-vectors/cloud-computing.svg',
-          slides:
-            'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-          code: 'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-        },
-        {
-          title: 'Cloud Computing 5',
-          description: 'Description',
-          emoji: '/assets/design-vectors/cloud-computing.svg',
-          slides:
-            'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-          code: 'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-        },
-        {
-          title: 'Cloud Computing 6',
-          description: 'Description',
-          emoji: '/assets/design-vectors/cloud-computing.svg',
-          slides:
-            'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-          code: 'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-        },
-        {
-          title: 'Cloud Computing 7',
-          description: 'Description',
-          emoji: '/assets/design-vectors/cloud-computing.svg',
-          slides:
-            'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-          code: 'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-        },
-        {
-          title: 'Cloud Computing 8',
-          description: 'Description',
-          emoji: '/assets/design-vectors/cloud-computing.svg',
-          slides:
-            'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-          code: 'https://docs.google.com/presentation/d/11oftaFBBf0rNHqF1MPV_dz0yyrLq_MW-AKh5RjDo1zE/edit?slide=id.p#slide=id.p',
-        },
-      ],
-    },
+export default function PastWorkshops() {
+  const topics = [
+    'Git',
+    'Cloud Computing',
+    'Career Prep',
+    'Web Development',
+    'Data Science',
+    'Machine Learning',
+    'Cybersecurity',
   ];
 
-  const [chunkSize, setChunkSize] = useState(3);
+  const [currentIndex, setCurrentIndex] = useState(0);
+  const TOPICS_PER_VIEW = 4;
 
-  useEffect(() => {
-    const handleResize = () => {
-      const width = window.innerWidth;
-      if (width <= 768) {
-        setChunkSize(1); // mobile
-      } else if (width <= 1200) {
-        setChunkSize(2); // tablet
-      } else setChunkSize(3); // desktop
-    };
+  const showArrows = topics.length > TOPICS_PER_VIEW;
+  const canScrollLeft = currentIndex > 0;
+  const canScrollRight = currentIndex + TOPICS_PER_VIEW < topics.length;
 
-    handleResize();
-    window.addEventListener('resize', handleResize);
-    return () => window.removeEventListener('resize', handleResize);
-  }, []);
-
-  // Helper function to chunk the workshops into groups of a specified size
-  const chunkWorkshops = (arr, size) => {
-    const chunks = [];
-    for (let i = 0; i < arr.length; i += size) {
-      chunks.push(arr.slice(i, i + size));
+  const scrollLeft = () => {
+    if (canScrollLeft) {
+      setCurrentIndex(Math.max(0, currentIndex - TOPICS_PER_VIEW));
     }
-    return chunks;
+  };
+
+  const scrollRight = () => {
+    if (canScrollRight) {
+      const maxIndex = Math.max(0, topics.length - TOPICS_PER_VIEW);
+      setCurrentIndex(Math.min(maxIndex, currentIndex + TOPICS_PER_VIEW));
+    }
   };
 
   return (
-    <div className={styles.container}>
-      {data.map((series, seriesIndex) => {
-        const rows = chunkWorkshops(series.workshops, chunkSize);
-
-        return (
-          <div key={seriesIndex} className={styles.seriesWrapper}>
-            {rows.map((row, rowIndex) => (
-              <div key={rowIndex} className={styles.rowWrapper}>
-                <div
-                  className={`${styles.horizontalLine} `}
-                  style={{
-                    width: `${row.length * 300 + (row.length - 1) * 20}px`,
-                    left: `${rowIndex % 2 === 0 ? '0' : 'auto'}`,
-                  }}
-                />
-
-                <div
-                  className={`${styles.row} ${rowIndex % 2 ? styles.odd : ''}`}
-                >
-                  {row.map((workshop, i) => (
-                    <WorkshopWindow
-                      key={i}
-                      className={styles.workshop}
-                      emoji={workshop.emoji}
-                      topic={workshop.title}
-                      slides={workshop.slides}
-                      code={workshop.code}
-                      recording={workshop.recording}
-                    />
-                  ))}
-                </div>
-
-                {rowIndex < rows.length - 1 && (
-                  <div
-                    className={`${styles.verticalLine} ${
-                      rowIndex % 2 === 0 ? styles.rightSide : styles.leftSide
-                    }`}
-                  />
-                )}
-              </div>
-            ))}
+    <div>
+      <div className={styles.container}>
+        <div className={styles.topicsWrapper}>
+          {showArrows && (
+            <button
+              type="button"
+              className={styles.arrowButton}
+              onClick={scrollLeft}
+              disabled={!canScrollLeft}
+              style={{ opacity: canScrollLeft ? 1 : 0.3 }}
+            >
+              <FaChevronLeft size={35} />
+            </button>
+          )}
+          <div className={styles.topicsContainer}>
+            <div
+              className={styles.topicsTrack}
+              style={{
+                transform: `translateX(-${currentIndex * (222 + 24)}px)`,
+              }}
+            >
+              {topics.map((topic, index) => (
+                <button type="button" className={styles.topic} key={index}>
+                  {topic}
+                </button>
+              ))}
+            </div>
           </div>
-        );
-      })}
+          {showArrows && (
+            <button
+              type="button"
+              className={styles.arrowButton}
+              onClick={scrollRight}
+              disabled={!canScrollRight}
+              style={{ opacity: canScrollRight ? 1 : 0.3 }}
+            >
+              <FaChevronRight size={35} />
+            </button>
+          )}
+        </div>
+      </div>
+      <div className={styles.seriesContainer}>
+        <WorkshopSeries />
+      </div>
     </div>
   );
 }
